+++ packages/iris-grid/src/LazyIrisGrid.tsx
@@ -1,8 +1,7 @@
-import { forwardRef, lazy, Suspense, useContext } from 'react';
+import { forwardRef, lazy, Suspense } from 'react';
 import { LoadingOverlay } from '@deephaven/components';
 import type IrisGridType from './IrisGrid';
 import type { IrisGridProps } from './IrisGrid';
-import CellInputRendererContext from './CellInputRendererContext';
 
 const IrisGrid = lazy(() => import('./IrisGrid.js'));
 
@@ -14,19 +13,14 @@ const LazyIrisGrid = forwardRef<
     // This creates the correct type to make defaultProps optional
     props,
     ref
-  ): JSX.Element => {
-    // Merge context-derived registry as default; caller-provided prop in {...props} wins
-    const registry = useContext(CellInputRendererContext);
-    const mergedProps = { cellInputRendererRegistry: registry, ...props };
-    return (
-      <Suspense
-        fallback={<LoadingOverlay data-testid="lazy-iris-grid-loading" />}
-      >
-        {/* eslint-disable-next-line react/jsx-props-no-spreading */}
-        <IrisGrid ref={ref} {...mergedProps} />
-      </Suspense>
-    );
-  }
+  ): JSX.Element => (
+    <Suspense
+      fallback={<LoadingOverlay data-testid="lazy-iris-grid-loading" />}
+    >
+      {/* eslint-disable-next-line react/jsx-props-no-spreading */}
+      <IrisGrid ref={ref} {...props} />
+    </Suspense>
+  )
 );
 
 LazyIrisGrid.displayName = 'LazyIrisGrid';
