+++ packages/iris-grid/src/index.ts
@@ -24,7 +24,7 @@ export * from './IrisGridTableModelTemplate';
 export { default as IrisGridModelFactory } from './IrisGridModelFactory';
 export { createDefaultIrisGridTheme } from './IrisGridTheme';
 export type { IrisGridThemeType } from './IrisGridTheme';
-export * from './IrisGridThemeProvider';
+export * from './IrisGridContextProvider';
 export { default as IrisGridTestUtils } from './IrisGridTestUtils';
 export { default as IrisGridUtils } from './IrisGridUtils';
 export * from './IrisGridUtils';
